ALL JOIN changes seems good execpt JOIN 0

+++ server.hpp
@@ -117,7 +117,7 @@ private:
     std::map<int, std::pair<std::string, std::string> >     userInfo_;
 
     Client		&findClient(std::string name);
-    void        joinChannel(Client& client, const std::vector<std::string>& params, int j);
+    void        joinChannel(Client& client, const std::vector<std::string>& params, size_t j);
     void        join(Channel *chan, Client &cl);
 
     typedef void (Server::*CommandHandler)(Client&, const std::vector<std::string>&);
